Rename StarCraftImageFile, add test
